refactor(cloudflare): move inferZoneIdForHostname beside resolveZoneId
The wrapper is not Worker-specific — it is `resolveZoneId` against the
ambient account with a per-run memo. Move it (and its tests) into
Zone/lookup.ts next to the walk it delegates to.
