+++ packages/alchemy/src/Cloudflare/Zone/lookup.ts
@@ -6,6 +6,7 @@ import * as zones from "@distilled.cloud/cloudflare/zones";
 import * as Effect from "effect/Effect";
 import * as Stream from "effect/Stream";
 import type * as HttpClient from "effect/unstable/http/HttpClient";
+import { CloudflareEnvironment } from "../CloudflareEnvironment.ts";
 
 /**
  * Reference to an existing Cloudflare Zone. Accepts:
@@ -44,6 +45,45 @@ export const resolveZoneId = ({
     return yield* Effect.fail(
       new Error(`Cloudflare zone not found for ${lookup}`),
     );
+  });
+
+/**
+ * {@link resolveZoneId} against the ambient account, for callers that only
+ * have a hostname: no explicit zone reference, a per-run memo, and an
+ * unresolvable hostname as a defect rather than a typed failure.
+ *
+ * Listing the account's zones and matching locally is a trap: the list
+ * endpoint paginates (20 zones per page), so a hostname whose zone sits on a
+ * later page silently fails to resolve. `resolveZoneId` walks the label
+ * hierarchy with exact `?name=` lookups instead, which makes the account's
+ * zone count irrelevant.
+ *
+ * `zoneCache` is the caller's memo — one map per reconcile pass, so a Worker
+ * with several domains or routes in the same zone looks it up once.
+ */
+export const inferZoneIdForHostname = (
+  hostname: string,
+  zoneCache: Map<string, string>,
+) =>
+  Effect.gen(function* () {
+    const cached = zoneCache.get(hostname);
+    if (cached) return cached;
+
+    const { accountId } = yield* yield* CloudflareEnvironment;
+    const zoneId = yield* resolveZoneId({
+      accountId,
+      zone: undefined,
+      hostname,
+    }).pipe(
+      Effect.catch(() =>
+        Effect.die(
+          `Could not infer Cloudflare Zone for hostname "${hostname}". ` +
+            "Ensure the parent zone exists in this account.",
+        ),
+      ),
+    );
+    zoneCache.set(hostname, zoneId);
+    return zoneId;
   });
 
 type ZoneListItem = {
